requirements.txt
Django==1.9.13
django-crispy-forms==1.6.1
django-formtools==1.0
django-model-utils==3.0.0
django-notifications-hq==1.3
django-reversion==2.0.9
django-simple-captcha==0.5.5
djangorestframework==3.6.3
future==0.15.2
httplib2==0.9.2
jsonfield==2.0.2
olefile==0.44
Pillow==4.1.1
PyInstaller==3.2.1
PyMySQL==0.7.11
pypiwin32==219
pytz==2017.2
six==1.10.0
